+++ src/main/kotlin/calculator/application/calculator/factory/CalculatorFactory.kt
@@ -1,5 +1,6 @@
-package calculator.application.calculator
+package calculator.application.calculator.factory
 
+import calculator.application.calculator.Calculator
 import calculator.application.calculator.impl.StandardCalculator
 import calculator.domain.operation.AdditionBinaryOperation
 import calculator.domain.operation.BinaryOperation
