+++ src/main/kotlin/calculator/application/calculator/impl/StandardCalculator.kt
@@ -1,7 +1,7 @@
 package calculator.application.calculator.impl
 
 import calculator.application.calculator.Calculator
-import calculator.application.model.PositiveInteger
+import calculator.common.PositiveInteger
 import calculator.domain.operation.BinaryOperation
 import calculator.domain.operation.BinaryOperationCommand
