+++ src/main/kotlin/calculator/assemble/ObjectAssemble.kt
@@ -1,10 +1,11 @@
 package calculator.assemble
 
 import calculator.application.calculator.Calculator
-import calculator.application.calculator.CalculatorFactory
+import calculator.application.calculator.factory.CalculatorFactory
 import calculator.application.parser.DelimiterParser
 import calculator.application.parser.Parser
 import calculator.interfaces.adapter.CalculatorViewAdapter
+import calculator.interfaces.adapter.ViewAdapter
 import calculator.interfaces.input.InputConsole
 import calculator.interfaces.input.InputPlugin
 import calculator.interfaces.output.OutputConsole
@@ -12,7 +13,7 @@ import calculator.interfaces.output.OutputPlugin
 
 object ObjectAssemble {
 
-    fun viewAdapter(): CalculatorViewAdapter = CalculatorViewAdapter(
+    fun viewAdapter(): ViewAdapter = CalculatorViewAdapter(
         inputPlugin = inputPlugin(),
         outputPlugin = outputPlugin(),
         calculator = calculator(),
