+++ src/main/kotlin/calculator/interfaces/adapter/CalculatorViewAdapter.kt
@@ -1,9 +1,9 @@
 package calculator.interfaces.adapter
 
 import calculator.application.calculator.Calculator
-import calculator.application.model.PositiveInteger
 import calculator.application.parser.Parser
 import calculator.application.parser.ParsingException
+import calculator.common.PositiveInteger
 import calculator.interfaces.input.InputPlugin
 import calculator.interfaces.output.OutputPlugin
 
@@ -12,9 +12,9 @@ class CalculatorViewAdapter(
     private val outputPlugin: OutputPlugin,
     private val calculator: Calculator,
     private val parser: Parser
-) {
+) : ViewAdapter {
 
-    fun start() {
+    override fun start() {
         val numberList = queryNumber()
         val result = calculator.multiplePlus(numberList)
         outputPlugin.output("결과: $result")
