+++ src/main/kotlin/calculator/interfaces/adapter/ViewAdapter.kt
@@ -1,0 +1,5 @@
+package calculator.interfaces.adapter
+
+sealed interface ViewAdapter {
+    fun start()
+}
